refactor(think): plan-section.md を template と SKILL.md 統合に分解
- templates/plan.md 新設。issue/templates と同じ書き方 (テンプレート節 +
  {...} placeholder + ガイドライン表) で骨格・id 記法・行数規則・抽出安定性
  を持つ
- contract / 前提の authoring 規則と書き出し前検証は think SKILL.md 本文へ
  統合。Phase 3 が毎回読む規則なので reference の 1 hop を消す
- references/plan-section.md は削除、テストは plan-draft.test.js に改名して
  template と SKILL.md の 2 ターゲットをピン
- build.js のコメント参照を templates/plan.md へ更新
